+++ packages/amplify-category-analytics/src/analytics-resource-api.ts
@@ -163,6 +163,7 @@ export const analyticsPushYes = async (context: $TSContext): Promise<void> => {
     context.exeInfo.inputParams = context.exeInfo.inputParams || {};
     context.exeInfo.inputParams.yes = true; // force yes to avoid prompts
     context.parameters = context.parameters || {};
+    context.parameters.options = context.parameters.options ?? {};
     context.parameters.options.yes = true;
     context.parameters.first = undefined;
     await analyticsPush(context);