+++ packages/amplify-category-auth/src/__tests__/provider-utils/awscloudformation/utils/auth-defaults-appliers.test.ts
@@ -1,4 +1,5 @@
-import { $TSAny, $TSContext } from 'amplify-cli-core';
+import { $TSAny } from 'amplify-cli-core';
+import { AuthContext } from '../../../../context';
 import { ServiceQuestionHeadlessResult } from '../../../../provider-utils/awscloudformation/service-walkthrough-types/cognito-user-input-types';
 import { structureOAuthMetadata } from '../../../../provider-utils/awscloudformation/service-walkthroughs/auth-questions';
 import {
@@ -43,7 +44,7 @@ describe('update auth defaults applier', () => {
       authSelections: 'userPoolOnly',
     } as ServiceQuestionHeadlessResult;
 
-    const result = await getUpdateAuthDefaultsApplier({} as unknown as $TSContext, 'cognito-defaults.js', {} as $TSAny)(stubResult);
+    const result = await getUpdateAuthDefaultsApplier({} as unknown as AuthContext, 'cognito-defaults.js', {} as $TSAny)(stubResult);
     expect(result).toMatchSnapshot();
     expect(structureOAuthMetadata_mock.mock.calls.length).toBe(1);
   });
@@ -55,7 +56,7 @@ describe('update auth defaults applier', () => {
       requiredAttributes: [] as string[],
     } as ServiceQuestionHeadlessResult;
 
-    const result = await getUpdateAuthDefaultsApplier({} as unknown as $TSContext, 'cognito-defaults.js', {} as $TSAny)(stubResult);
+    const result = await getUpdateAuthDefaultsApplier({} as unknown as AuthContext, 'cognito-defaults.js', {} as $TSAny)(stubResult);
     expect(result.requiredAttributes).toEqual([]);
   });
 });
@@ -68,7 +69,7 @@ describe('add auth defaults applier', () => {
       requiredAttributes: [] as string[],
     } as ServiceQuestionHeadlessResult;
 
-    const result = await getAddAuthDefaultsApplier({} as unknown as $TSContext, 'cognito-defaults.js', 'testProjectName')(stubResult);
+    const result = await getAddAuthDefaultsApplier({} as unknown as AuthContext, 'cognito-defaults.js', 'testProjectName')(stubResult);
     expect(result.requiredAttributes).toEqual([]);
   });
 });