+++ packages/amplify-cli/src/__tests__/commands/init.test.ts
@@ -118,7 +118,7 @@ describe('amplify init:', () => {
           },
         },
       };
-      await preInitSetup(context);
+      await preInitSetup(context as unknown as $TSContext);
       expect(execSync).toBeCalledWith(`git ls-remote ${appUrl}`, { stdio: 'ignore' });
       expect(execSync).toBeCalledWith(`git clone ${appUrl} .`, { stdio: 'inherit' });
       expect(execSync).toBeCalledWith('yarn install', { stdio: 'inherit' });