+++ packages/amplify-cli/src/domain/amplify-usageData/identifiable-input-regex.ts
@@ -1,5 +1,5 @@
 import { JSONUtilities } from 'amplify-cli-core';
-import { CommandLineInput } from 'amplify-cli-core';
+import { CLIInput } from '../command-input';
 
 const containsToRedact = ['key', 'id', 'password', 'name', 'arn', 'address', 'app'];
 const quotes = '\\\\?"';
@@ -15,7 +15,7 @@ function testReplaceJsonValues(json: string, redactedInput: string): string {
   if (jsonRegex.test(s)) {
     jsonRegex.lastIndex = 0;
     let m: RegExpExecArray | null;
-    const valuesToRedact: any = [];
+    const valuesToRedact: string[] = [];
 
     //find all values to redact
     do {
@@ -35,12 +35,8 @@ function testReplaceJsonValues(json: string, redactedInput: string): string {
   return s;
 }
 
-export default function redactInput(
-  originalInput: CommandLineInput,
-  deleteArgAndOption: boolean,
-  replacementString = '************',
-): CommandLineInput {
-  const input: CommandLineInput = JSONUtilities.parse(JSONUtilities.stringify(originalInput)!);
+export default function redactInput(originalInput: CLIInput, deleteArgAndOption: boolean, replacementString = '************'): CLIInput {
+  const input: CLIInput = JSONUtilities.parse(JSONUtilities.stringify(originalInput) as string);
   const argv = input.argv;
   const length = argv.length;
   let redactString = false;