+++ packages/amplify-category-function/CHANGELOG.md
@@ -1001,6 +1001,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [4.3.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@4.3.1...amplify-category-function@4.3.2) (2023-02-16)
+
+**Note:** Version bump only for package amplify-category-function
+
+
+
+
+
 ## [4.3.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-function@4.3.0...amplify-category-function@4.3.1) (2023-02-06)
 
 **Note:** Version bump only for package amplify-category-function