+++ packages/amplify-category-storage/CHANGELOG.md
@@ -214,6 +214,14 @@ Co-authored-by: Akshay Upadhyay <akz@amazon.com>
 
 
 ## [4.0.1-beta.2](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-storage@4.0.1-beta.1...@aws-amplify/amplify-category-storage@4.0.1-beta.2) (2022-11-19)
+## [3.7.2](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-storage@3.7.1...@aws-amplify/amplify-category-storage@3.7.2) (2023-02-16)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-storage
+
+
+
+
+
 ## [3.7.1](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-category-storage@3.7.0...@aws-amplify/amplify-category-storage@3.7.1) (2023-02-06)
 
 **Note:** Version bump only for package @aws-amplify/amplify-category-storage