+++ packages/amplify-nodejs-function-template-provider/CHANGELOG.md
@@ -249,6 +249,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [2.6.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@2.6.4...amplify-nodejs-function-template-provider@2.6.5) (2023-02-16)
+
+**Note:** Version bump only for package amplify-nodejs-function-template-provider
+
+
+
+
+
 ## [2.6.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-nodejs-function-template-provider@2.6.3...amplify-nodejs-function-template-provider@2.6.4) (2023-02-06)
 
 **Note:** Version bump only for package amplify-nodejs-function-template-provider